requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.2.5
django-bootstrap4==3.0.1
django-stdimage==5.3.0
gunicorn==20.1.0
mysql==0.0.3
mysqlclient==2.0.3
Pillow==8.3.1
psycopg2-binary==2.9.1
PyMySQL==1.0.2
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
whitenoise==5.3.0
